+++ src/EventListener/CheckoutConfirmTrustlyEventListener.php
@@ -82,7 +82,7 @@ class CheckoutConfirmTrustlyEventListener implements EventSubscriberInterface
 
         $billingAddress = $customer->getActiveBillingAddress();
 
-        if (null === $billingAddress) {
+        if (null === $billingAddress || null === $billingAddress->getCountry()) {
             return false;
         }