+++ tests/Payone/Request/CreditCard/CreditCardPreAuthorizeRequestFactoryTest.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace PayonePayment\Test\Payone\Request\CreditCard;
 
 use DMS\PHPUnitExtensions\ArraySubset\Assert;
+use PayonePayment\Components\ConfigReader\ConfigReader;
 use PayonePayment\Components\RedirectHandler\RedirectHandler;
+use PayonePayment\Configuration\ConfigurationPrefixes;
 use PayonePayment\Installer\CustomFieldInstaller;
 use PayonePayment\PaymentHandler\PayoneCreditCardPaymentHandler;
 use PayonePayment\Payone\Request\CreditCard\CreditCardPreAuthorizeRequest;
 use PayonePayment\Payone\Request\CreditCard\CreditCardPreAuthorizeRequestFactory;
+use PayonePayment\Struct\Configuration;
 use PayonePayment\Struct\PaymentTransaction;
 use PayonePayment\Test\Constants;
 use PayonePayment\Test\Mock\Factory\RequestFactoryTestTrait;
@@ -45,13 +48,11 @@ class CreditCardPreAuthorizeRequestFactoryTest extends TestCase
                 'aid'             => '',
                 'amount'          => 10000,
                 'api_version'     => '3.10',
-                'backurl'         => '',
                 'city'            => 'Some City',
                 'clearingtype'    => 'cc',
                 'currency'        => 'EUR',
                 'email'           => 'first.last@example.com',
                 'encoding'        => 'UTF-8',
-                'errorurl'        => '',
                 'firstname'       => 'First',
                 'integrator_name' => 'shopware6',
                 'key'             => '',
@@ -65,7 +66,6 @@ class CreditCardPreAuthorizeRequestFactoryTest extends TestCase
                 'request'         => 'preauthorization',
                 'solution_name'   => 'kellerkinder',
                 'street'          => 'Some Street 1',
-                'successurl'      => '',
                 'zip'             => '12345',
             ],
             $request
@@ -122,6 +122,13 @@ class CreditCardPreAuthorizeRequestFactoryTest extends TestCase
             )
         );
 
-        return new CreditCardPreAuthorizeRequest($this->createMock(RedirectHandler::class), $currencyRepository);
+        $configReader = $this->createMock(ConfigReader::class);
+        $configReader->method('read')->willReturn(
+            new Configuration([
+                sprintf('%sProvideNarrativeText', ConfigurationPrefixes::CONFIGURATION_PREFIX_CREDITCARD) => false,
+            ])
+        );
+
+        return new CreditCardPreAuthorizeRequest($this->createMock(RedirectHandler::class), $currencyRepository, $configReader);
     }
 }